requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
bcrypt==4.0.1
billiard==4.2.4
bitarray==3.8.0
celery==5.3.6
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==1.0.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cytoolz==1.1.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.1
fastapi==0.109.0
frozenlist==1.8.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
langchain==0.1.4
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.5
langsmith==0.0.87
lru-dict==1.2.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.10.0
packaging==23.2
parsimonious==0.10.0
passlib==1.7.4
pinecone-client==3.0.2
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.9
py-ecc==8.0.0
pycryptodome==3.23.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.10.1
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.7
pyunormalize==17.0.0
pyngrok==7.5.0
PyYAML==6.0.3
redis==5.0.1
referencing==0.36.2
regex==2025.11.3
requests==2.31.0
rlp==4.1.0
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
stripe==7.9.0
tenacity==8.5.0
tiktoken==0.5.2
tomli==2.3.0
toolz==1.1.0
tqdm==4.67.1
twilio==8.11.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.27.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
